backend/requirements.txt
fastapi==0.115.8
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
uvicorn==0.34.0